Refactor: Rename `email` module to `email_service` and update imports across the codebase
