+++ src/core/email_service/config.py
