+++ src/core/email_service/interfaces.py
