+++ src/core/email_service/schemas.py
