+++ src/core/email_service/service.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from fastapi_mail import MessageType
 from pydantic import EmailStr, ValidationError, TypeAdapter, BaseModel
 
-from src.core.email.interfaces import AbstractMailer
+from src.core.email_service.interfaces import AbstractMailer
 from loggers import get_logger
 
 logger = get_logger(__name__)
