+++ src/core/email_service/templates/verification.html
